+++ R/ggpairs.R
@@ -618,6 +618,7 @@ ggduo <- function(
 #' @template ggmatrix-legend-param
 #' @param cardinality_threshold maximum number of levels allowed in a charcter / factor column.  Set this value to NULL to not check factor columns. Defaults to 15
 #' @param legends deprecated
+#' @param labeller labeller for facets. See \code{\link[ggplot2]{labellers}}. Common values are \code{"label_value"} (default) and \code{"label_parsed"}.
 #' @keywords hplot
 #' @import ggplot2
 #' @references John W Emerson, Walton A Green, Barret Schloerke, Jason Crowley, Dianne Cook, Heike Hofmann, Hadley Wickham. The Generalized Pairs Plot. Journal of Computational and Graphical Statistics, vol. 22, no. 1, pp. 79-91, 2012.
@@ -681,6 +682,13 @@ ggduo <- function(
 #' pm <- ggpairs(tips[, 1:3], axisLabels="none")
 #' p_(pm)
 #'
+#' ## Facet Label Variations
+#' #  Default:
+#' df <- tibble::data_frame(x = rnorm(100), y = a + rnorm(100, 0, 0.1), c = sqrt(x^2 +  y^2))
+#' ggpairs(df, columnLabels = c("alpha[foo]", "alpha[bar]", "sqrt(alpha[foo]^2 + alpha[bar]^2)"))
+#' #  Parsed labels:
+#' ggpairs(df, columnLabels = c("alpha[foo]", "alpha[bar]", "sqrt(alpha[foo]^2 + alpha[bar]^2)"), labeller = "label_parsed")
+#'
 #' ## Plot Insertion Example
 #' custom_car <- ggpairs(mtcars[, c("mpg", "wt", "cyl")], upper = "blank", title = "Custom Example")
 #' # ggplot example taken from example(geom_text)
@@ -722,7 +730,8 @@ ggpairs <- function(
   showStrips = NULL,
   legend = NULL,
   cardinality_threshold = 15,
-  legends = stop("deprecated")
+  legends = stop("deprecated"),
+  labeller = 'label_value'
 ){
 
   warn_deprecated(!missing(legends), "legends")
@@ -817,7 +826,8 @@ ggpairs <- function(
     ylab = ylab,
     data = data,
     gg = NULL,
-    legend = legend
+    legend = legend,
+    labeller = labeller
   )
 
   plotMatrix